Separate common FDTD implementation into a base class
